hot bug fixed

+++ pyod/models/sod.py
@@ -30,11 +30,11 @@ def _snn_imp(ind, ref_set_):
     n = ind.shape[0]
     _count = np.zeros(shape=(n, ref_set_), dtype=nb.uint32)
     for i in nb.prange(n):
-        temp = np.empty(n, dtype=nb.int16)
+        temp = np.empty(n, dtype=nb.uint32)
         test_element_set = set(ind[i])
         for j in nb.prange(n):
             temp[j] = len(set(ind[j]).intersection(test_element_set))
-        temp[i] = np.iinfo(np.uint16).max
+        temp[i] = np.iinfo(np.uint32).max
         _count[i] = np.argsort(temp)[::-1][1:ref_set_ + 1]
 
     return _count
